Add tentative code to calibration
No breaking changes, but in the case of faulty connections, the tubing
is filled with air bubbles leading to inconsistent bolus sizes.  It
might make sense to trigger the vials in sequence rather than at the
same event.  Exploring this idea by storing a new log for calibration,
which tracks the last vial for which the influx has been triggered.
TODO: spin off a new branch for this function, test, and ask for feedback