Populate setup page with current yaml settings
Reads the current experiment_parameters.yaml if it exists and
populates the default values from this file.

Additionally, execute the write_parameters_to_file.py script from the
browser once the calibration ends. In the future, this script will be
integrated with the visualize.py script to avoid system calls.