Fixed bug in transform_data
Original code assumed that the first dataset in the od config
corresponded to OD90 and the second to OD135 if at all. For some
reason this wasn't true on Mermaidman, and the returned values were
nonsensical. I have now added a check to read the key of the config
variable and test it for the index of OD90.